Collapse intent code into abilities module
