+++ emergence_lib/src/player_interaction/mod.rs
@@ -10,7 +10,6 @@ use leafwing_input_manager::{
 pub(crate) mod abilities;
 pub(crate) mod camera;
 pub(crate) mod clipboard;
-pub(crate) mod intent;
 pub(crate) mod picking;
 pub(crate) mod selection;
 
@@ -25,7 +24,6 @@ impl Plugin for InteractionPlugin {
             .add_plugin(camera::CameraPlugin)
             .add_plugin(abilities::AbilitiesPlugin)
             .add_plugin(picking::PickingPlugin)
-            .add_plugin(intent::IntentPlugin)
             .add_plugin(selection::SelectionPlugin)
             .add_plugin(clipboard::ClipboardPlugin);
 
@@ -45,12 +43,8 @@ pub(crate) enum InteractionSystem {
     SelectTiles,
     /// Held structure(s) are selected
     SetClipboard,
-    /// Replenishes the [`IntentPool`](intent::IntentPool) of the hive mind
-    ReplenishIntent,
     /// Apply zoning to tiles
     ApplyZoning,
-    /// Use intent-spending abilities
-    UseAbilities,
     /// Spawn and despawn ghosts
     ManagePreviews,
 }
